refactor: Lotto 당첨 번호 로직 리팩토링
- LottoNumbers 클래스 생성 해서 내부에서 검증 하는 방식으로 변경
