+++ src/test/java/lotto/README.md
@@ -5,9 +5,6 @@
 ### JavaLotto(총괄 컨트롤러)
 - UI 로직 포함이라 TDD X
 
-### Lotto(로또 1장)
-- 당첨자 번호와 매치 몇개 맞았는가? 검증
-
 ### LottoMachine(로또 기계 - 자동 번호 발급, 당첨 리스트 반환) 
 - 금액이 입력되었을 때 1000원으로 나누었을 때 발급된 로또 갯수가 몇 장인가?
 - 발급 리스트와 당첨 번호 리스트 입력 시 당첨 리스트 반환 
@@ -30,3 +27,6 @@
 
 ### Lotto(로또 1장)
 - 로또 1장 정렬 검증
+- 당첨 번호와 매치 시 몇 개 맞았는가? 검증
+
+### LottoNumbers(로또 1장의 번호 리스트)
